Force header / footer nodes format.
get_headers_footers_xml was returning the xml in a raw string without any
processing on it. If we had a quote in docx document, it was encode in a xml
entity (&quot;), but because of the entities, Jinja raise error.

Now, xml passes through etree.tostring and we have a clean string.